requirements.txt
edge-tts
pydub
flask
PyPDF2
google-genai
